Replace NanoLimboPluginHook with NanoLimboProvider
