+++ velocity/src/main/java/me/mastercapexd/auth/velocity/hooks/nanolimbo/VelocityNanoLimboPluginHook.java
@@ -1,64 +1,35 @@
 package me.mastercapexd.auth.velocity.hooks.nanolimbo;
 
 import java.net.InetSocketAddress;
-import java.nio.charset.StandardCharsets;
-import java.util.Collections;
 import java.util.stream.IntStream;
 
+import com.bivashy.auth.api.hook.LimboPluginHook;
 import com.velocitypowered.api.proxy.ProxyServer;
 import com.velocitypowered.api.proxy.server.ServerInfo;
-import com.velocitypowered.proxy.config.PlayerInfoForwarding;
-import com.velocitypowered.proxy.config.VelocityConfiguration;
 
-import me.mastercapexd.auth.hooks.nanolimbo.NanoLimboPluginHook;
+import me.mastercapexd.auth.hooks.nanolimbo.NanoLimboProvider;
 import me.mastercapexd.auth.velocity.server.VelocityProxyServer;
-import ua.nanit.limbo.NanoLimbo;
-import ua.nanit.limbo.server.LimboServer;
-import ua.nanit.limbo.server.data.InfoForwarding;
 
-public class VelocityNanoLimboPluginHook implements NanoLimboPluginHook {
+public class VelocityNanoLimboPluginHook implements LimboPluginHook {
 
     private final IntStream limboPortRange;
-    private final ClassLoader classLoader;
     private final ProxyServer proxyServer;
+    private NanoLimboProvider provider;
 
-    public VelocityNanoLimboPluginHook(IntStream limboPortRange, ProxyServer proxyServer, String portRange) {
+    public VelocityNanoLimboPluginHook(IntStream limboPortRange, ProxyServer proxyServer) {
         this.limboPortRange = limboPortRange;
-        this.classLoader = NanoLimbo.class.getClassLoader();
         this.proxyServer = proxyServer;
+        if (!canHook())
+            return;
+        provider = new VelocityNanoLimboProvider(getClass().getClassLoader(), proxyServer);
     }
 
     @Override
     public com.bivashy.auth.api.server.proxy.ProxyServer createServer(String serverName) {
-        InetSocketAddress address = findAvailableAddress(limboPortRange).orElseThrow(
+        InetSocketAddress address = provider.findAvailableAddress(limboPortRange).orElseThrow(
                 () -> new IllegalStateException("Cannot find available port for limbo server!"));
-        LimboServer server = createLimboServer(address);
-        try {
-            server.start();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
+        provider.createAndStartLimbo(address);
         return new VelocityProxyServer(proxyServer.registerServer(new ServerInfo(serverName, address)));
-    }
-
-    @Override
-    public InfoForwarding createForwarding() {
-        VelocityConfiguration velocityConfiguration = (VelocityConfiguration) proxyServer.getConfiguration();
-        PlayerInfoForwarding forwardingMode = velocityConfiguration.getPlayerInfoForwardingMode();
-        if (forwardingMode == PlayerInfoForwarding.NONE)
-            return FORWARDING_FACTORY.none();
-        if (forwardingMode == PlayerInfoForwarding.LEGACY)
-            return FORWARDING_FACTORY.legacy();
-        if (forwardingMode == PlayerInfoForwarding.MODERN)
-            return FORWARDING_FACTORY.modern(velocityConfiguration.getForwardingSecret());
-        if (forwardingMode == PlayerInfoForwarding.BUNGEEGUARD)
-            return FORWARDING_FACTORY.bungeeGuard(Collections.singleton(new String(velocityConfiguration.getForwardingSecret(), StandardCharsets.UTF_8)));
-        return FORWARDING_FACTORY.none();
-    }
-
-    @Override
-    public ClassLoader classLoader() {
-        return classLoader;
     }
 
     @Override
